Divide Tokenizer tests into multi files
