+++ Tests/TokenizerTest/TokenizeErrorTest.swift
@@ -1,0 +1,13 @@
+import XCTest
+@testable import Tokenizer
+
+final class TokenizeErrorTest: XCTestCase {
+
+    func testFailUnknownToken() throws {
+        do {
+            _ = try tokenize(source: "1 ^")
+        } catch let error as TokenizeError {
+            XCTAssertEqual(error, TokenizeError.unknownToken(index: 2))
+        }
+    }
+}
